[#167347769] Update approach to modal on FiscalCode screen (#1228)
* remove commented code

* shift FiscalCodeLandscapeModal

* update reviewed modal name

* remove commented code

* remove nomore used screen state

* remove unused code

* upodate usage of platform

* update2
